refactor(internal/config): change width and height types to float32

+++ internal/config/gui.go
@@ -1,12 +1,12 @@
 package config
 
 type GUIConfig struct {
-	Width     int
-	Height    int
+	Width     float32
+	Height    float32
 	Resizable bool
 }
 
-func BuildGUIConfig(width, height int, resizable bool) *GUIConfig {
+func BuildGUIConfig(width, height float32, resizable bool) *GUIConfig {
 	return &GUIConfig{
 		Width:     width,
 		Height:    height,
